Build, start, and run scripts for all .NET targets
Some fixes to make sure that they all function correctly

+++ servers/dotnet/scripts/stop_winui.ps1
@@ -1,2 +1,10 @@
-$childPID = Get-Process -ProcessName testserver | Select-Object -Expand Id
+Import-Module $PSScriptRoot/prepare_env.psm1 -Force
+
+$childPID = Get-Process -ProcessName testserver -ErrorAction Ignore | Select-Object -Expand Id
+if($null -eq $childPID) {
+    Write-Error "No running testserver found!"
+    exit 1
+}
+
+Banner "Stopping testserver ($childPID)"
 Stop-Process $childPID
